Rename api/lib/settings.js into api/lib/config.js.
The convention selected is:
- word "settings" is about options managed by users
- word "config" is about options managed by developers
