+++ api/bin/www
@@ -3,7 +3,7 @@
 const heapProfile = require('heap-profile');
 const createServer = require('../server');
 const logger = require('../lib/infrastructure/logger');
-const { system } = require('../lib/settings');
+const { system } = require('../lib/config');
 
 const start = async () => {
   try {
