+++ api/tests/acceptance/application/organization-controller_test.js
@@ -2,7 +2,7 @@ const jwt = require('jsonwebtoken');
 const { expect, knex, nock, databaseBuilder, generateValidRequestAuthorizationHeader, insertUserWithRolePixMaster, cleanupUsersAndPixRolesTables } = require('../../test-helper');
 const Membership = require('../../../lib/domain/models/Membership');
 const createServer = require('../../../server');
-const settings = require('../../../lib/settings');
+const settings = require('../../../lib/config');
 const areaRawAirTableFixture = require('../../tooling/fixtures/infrastructure/areaRawAirTableFixture');
 const _ = require('lodash');
 const iconv = require('iconv-lite');
